pyproject.toml
[project]
name = "milstack-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.4.1",
    "gradio>=5.23.0",
    "haystack-ai>=2.11.2",
    "haystack-experimental>=0.8.0",
    "ollama-haystack>=2.3.0",
    "sentence-transformers>=3.4.1",
    "vllm>=0.8.1",
]
